Updates to restructure/consolidate some of the testing code
