+++ tests_new/performance/netmiko_performance.csv
@@ -22,3 +22,4 @@ date,netmiko_version,device_name,connect,send_command_simple,send_config_simple,
 2020-8-28 15:19:4,3.3.0_dev,cisco5,0:00:01.876103,0:00:01.442482,0:00:02.132264,0:00:11.050844
 2020-8-28 15:33:54,3.2.1_dev,nxos1,0:00:08.107724,0:00:08.487870,0:00:07.868704,0:00:17.714860
 2020-8-29 9:48:26,3.3.0_dev,nxos1,0:00:02.530558,0:00:02.931070,0:00:02.733514,0:00:12.587129
+2020-9-1 11:1:57,3.3.0,nxos1,0:00:02.234069,0:00:02.413287,0:00:02.830380,0:00:10.134148
